+++ Pandora/Paradigm/Structure/Ability/Focusable.hs
@@ -5,5 +5,5 @@ import Pandora.Paradigm.Inventory.Optics (type (:-.))
 
 class Focusable t where
 	type Focus (t :: * -> *) a
-	root :: t a :-. Focus t a
+	top :: t a :-. Focus t a
 	singleton :: a |-> t