+++ CHANGELOG.md
@@ -228,3 +228,4 @@
 * Move `Ordering` definition to its own module
 * Rename `Variation` submodule to `Ability` and create umbrella-module
 * Define `Focusable` typeclass for getting root and creating singleton
+* Replace `top` method of `Stack` structure with `Focusable` instance